Use XxHash64

+++ src/EFCoreSecondLevelCacheInterceptor/IEFHashProvider.cs
@@ -10,23 +10,23 @@ namespace EFCoreSecondLevelCacheInterceptor
         /// </summary>
         /// <param name="data">the input string</param>
         /// <returns>Hash</returns>
-        uint ComputeHash(string data);
+        ulong ComputeHash(string data);
 
         /// <summary>
-        /// Computes the Hash of the input byte array.
+        /// Computes the hash of the input array.
         /// </summary>
-        /// <param name="data">the input byte array</param>
+        /// <param name="data">the input array</param>
         /// <returns>Hash</returns>
-        uint ComputeHash(byte[] data);
+        ulong ComputeHash(byte[] data);
 
         /// <summary>
-        /// Computes the Hash of the input byte array.
+        /// Computes the hash of the input byte array.
         /// </summary>
         /// <param name="data">the input byte array</param>
         /// <param name="offset">start offset</param>
         /// <param name="len">length</param>
         /// <param name="seed">initial seed</param>
         /// <returns>Hash</returns>
-        uint ComputeHash(byte[] data, int offset, uint len, uint seed);
+        ulong ComputeHash(byte[] data, int offset, int len, uint seed);
     }
 }
